requirements.txt
flask==3.1.3
flask-cors==6.0.0
requests==2.32.4
anthropic==0.43.0
python-dotenv==1.0.1
feedparser==6.0.11
